requirements.txt
# requirements.txt

# --- Core Web Framework & Server ---
fastapi[all]
uvicorn[standard]

# --- Database ---
sqlmodel
psycopg2-binary
sqlalchemy

# --- Asynchronous HTTP Clients ---
httpx
aiohttp

# --- Background Task Processing ---
celery
redis

# --- AI & ML Models ---
# For Transcription
faster-whisper==1.1.1

# For Diarization & Speaker Enrollment (dependencies from your processing scripts)
# Note: The base image vint1-diarization:v1 likely has nemo and torch pre-installed.
# We list them here to ensure they are available if the base image is minimal.
nemo-toolkit[asr]
torch
librosa
soundfile
scikit-learn
silero-vad

# --- Vector Database Client ---
qdrant-client

# --- LLM Client & Utilities ---
openai>=1.3.0
python-dotenv

# --- Document Generation ---
pypandoc
python-docx

# --- Other Utilities ---
pydub
requests
Werkzeug